Moved type to new file
